Update dependencies and improve typing. Also change from webpack to esbuild.

+++ src/main.tsx
@@ -1,16 +1,18 @@
 import './style.scss';
-import { JSON, JsonObject, JsonProperty } from "ta-json";
-import { render, Component } from 'inferno';
+import * as React from 'react'
+import { JSON } from "ta-json";
 import { PreviewSign } from './view_preview';
 import { Sign } from './data';
 import "classcat"
 import { SettingsSign } from './view_settings';
 import { debounce } from 'ts-debounce';
+import { Component } from 'react';
+import { createRoot } from 'react-dom/client';
 
 function initializeWithJson(obj: any, json: any) {
     // TODO: Handle list
-    for(var prop in obj) {
-        if(!json.hasOwnProperty(prop)) {
+    for (var prop in obj) {
+        if (!json.hasOwnProperty(prop)) {
             continue;
         }
 
@@ -28,19 +30,23 @@ function initializeWithJson(obj: any, json: any) {
 
 function SignItem({ item, onOpen }: any) {
     return (
-        <li onClick={ onOpen }>
-            <span>{ item.name }</span>
+        <li onClick={onOpen}>
+            <span>{item.name}</span>
         </li>
     )
 }
 
-class SignSelector extends Component {
+interface SignSelectorProps {
+    selectedId: number | null;
+    onOpen: (id: null | number) => void;
+}
+class SignSelector extends Component<SignSelectorProps, { signs: { id: number, name: string }[] }> {
     constructor(props: any) {
         super(props);
         this.state = { signs: [] }
     }
 
-    componentWillMount() {
+    override componentWillMount() {
         this.findSigns();
     }
 
@@ -57,17 +63,17 @@ class SignSelector extends Component {
         });
     }
 
-    render() {
+    override render() {
         return (
-            <div class="sign-selector">
+            <div className="sign-selector">
                 <h1>MakerSign</h1>
                 <ul>
-                    <li onClick={ () => this.props.onOpen(null) }>
+                    <li onClick={() => this.props.onOpen(null)}>
                         <span>Create new sign</span>
                     </li>
                 </ul>
                 <ul>
-                    { this.state.signs.map((item: any) => SignItem({ item: item, onOpen: () => this.props.onOpen(item.id) })) }
+                    {this.state.signs.map((item: any) => SignItem({ item: item, onOpen: () => this.props.onOpen(item.id) }))}
                 </ul>
             </div>
         )
@@ -75,16 +81,16 @@ class SignSelector extends Component {
 }
 
 class SignSelectorSmall extends SignSelector {
-    componentDidUpdate(lastProps: any, lastState: any, snapshot: any) {
+    override componentDidUpdate(lastProps: SignSelectorProps, lastState: any, snapshot: any) {
         if (lastProps.selectedId != this.props.selectedId) {
             this.findSigns();
         }
     }
 
-    render() {
+    override render() {
         return (
-            <select class="sign-selector" value={this.props.selectedId != null ? this.props.selectedId : -1} onInput={e => { const id = Number((e.target as HTMLSelectElement).value); this.props.onOpen(id != -1 ? id : null); }}>
-                { this.state.signs.map((item: any) => (<option value={item.id}>{item.name}</option>)) }
+            <select className="sign-selector" value={this.props.selectedId != null ? this.props.selectedId : -1} onInput={e => { const id = Number((e.target as HTMLSelectElement).value); this.props.onOpen(id != -1 ? id : null); }}>
+                {this.state.signs.map(item => (<option key={item.id} value={item.id}>{item.name}</option>))}
                 <option value={-1}>Create new sign</option>
             </select>
         )
@@ -94,11 +100,11 @@ class SignSelectorSmall extends SignSelector {
 // How often to save (at most). In milliseconds
 const SavingInterval = 2000;
 
-export class App extends Component {
-    saving: boolean;
+export class App extends Component<{}, { id: number | null, sign: Sign | null }> {
+    saving: boolean = false;
     debouncedSave = debounce(() => this.save(), SavingInterval);
 
-    constructor(props: any) {
+    constructor(props: {}) {
         super(props);
         this.state = { sign: null, id: null };
         this.openFromURL();
@@ -110,7 +116,7 @@ export class App extends Component {
         this.debouncedSave();
     }
 
-    componentWillMount() {
+    override componentWillMount() {
         window.onpopstate = () => this.openFromURL();
         // this.openFromURL();
     }
@@ -124,7 +130,7 @@ export class App extends Component {
         }
     }
 
-    open(id: number|null, pushState: boolean = true) {
+    open(id: number | null, pushState: boolean = true) {
         if (this.state.id == id && id !== null) return;
 
         if (id === null) {
@@ -142,11 +148,11 @@ export class App extends Component {
                 const item = json.data;
                 const sign = new Sign();
                 initializeWithJson(sign, item.data);
-                this.setState({ sign: sign, id: item.id});
+                this.setState({ sign: sign, id: item.id });
                 if (pushState) history.pushState({}, "", `/${id}`);
             }).catch(e => {
                 console.error(e);
-                this.setState({ sign: null, id: null});
+                this.setState({ sign: null, id: null });
                 history.pushState({}, "", `/`);
             });
         }
@@ -177,14 +183,14 @@ export class App extends Component {
 
     }
 
-    render() {
-        if (this.state.sign != null) {
+    override render() {
+        if (this.state.sign != null && this.state.id !== null) {
             return (
-                <div class="app-root">
+                <div className="app-root">
                     <div id="settings">
-                        <div class="sign-root">
-                            <SignSelectorSmall selectedId={this.state.id} onOpen={ (id: number|null) => this.open(id) } />
-                            <SettingsSign sign={this.state.sign} onChange={ () => this.onChange() } onSave={ () => this.save() } onDelete={ () => this.delete() } autosaved={this.state.id !== null} />
+                        <div className="sign-root">
+                            <SignSelectorSmall selectedId={this.state.id} onOpen={(id: number | null) => this.open(id)} />
+                            <SettingsSign sign={this.state.sign} onChange={() => this.onChange()} onSave={() => this.save()} onDelete={() => this.delete()} autosaved={this.state.id !== null} />
                         </div>
                     </div>
                     <div id="preview">
@@ -194,16 +200,15 @@ export class App extends Component {
             );
         } else {
             return (
-                <div class="app-root">
-                    <SignSelector onOpen={ (id: number|null) => this.open(id) } />
+                <div className="app-root">
+                    <SignSelector selectedId={null} onOpen={(id: number | null) => this.open(id)} />
                 </div>
             )
         }
     }
 }
 
+const container = document.getElementById('app-root');
+const root = createRoot(container!);
+root.render(<App />);
 
-render(
-    <App />,
-    document.getElementById("app-root")
-);
